requirements.txt
astroid==1.4.5
colorama==0.3.7
coverage==3.7.1
Django==1.8.2
django-contact-form==1.0
django-coverage==1.2.4
django-crispy-forms==1.6.0
django-jenkins==0.18.1
django-localeurl==2.0.2
jslint==0.6.0
lazy-object-proxy==1.2.2
MySQL-python==1.2.5
pep8==1.7.0
Pillow==2.7.0
pycurl==7.43.0
pyflakes==1.1.0
pylint==1.5.5
python-dateutil==2.4.2
six==1.9.0
wrapt==1.10.8
